+++ anemone-kernel/src/fs/api/mod.rs
@@ -35,8 +35,8 @@ pub mod renameat2;
 pub mod sendfile;
 pub mod stat;
 pub mod statfs;
-pub mod truncate;
 pub mod symlinkat;
+pub mod truncate;
 pub mod umask;
 pub mod umount;
 pub mod unlinkat;
